requirements.txt
# ----- required
graphviz >= 0.14
gputil >= 1.4.0
matplotlib >= 3.2.1
numpy >= 1.18.4
torch >= 1.7.0
torchvision >= 0.8.0
pytorch-lightning >= 1.2.8
torchprofile >= 0.0.2

# ----- required, but should be made optional if possible
pymoo >= 0.4.2.1

# ----- optional, not installing will only make specific classes / functions unavailable
albumentations >= 0.5.2
nats_bench >= 1.1.0
scipy >= 1.5.1
scikit-learn >= 0.23.2
Pillow >= 7.1.0
adabound >= 0.0.5
adabelief-pytorch >= 0.0.5
pyro5 >= 5.11.0
timm >= 0.2.1
onnx >= 1.8.0
tensorboard >= 2.2.0
xgboost >= 1.4.0

# ----- optional, just testing around right now
mmcv-full >= 1.2.1
mmdet >= 2.6.0
pycocotools >= 2.0.2
# detectron2 >= 0.3
# nasbench301 >= 0.2
